backend/requirements.txt
# Core Django and REST Framework
Django==6.0
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg[binary]==3.2.3
python-dotenv==1.0.1
django-filter==24.3
Pillow==11.0.0

# Authentication
djangorestframework-simplejwt==5.3.1

# API Documentation
drf-spectacular==0.28.0

# Security
django-ratelimit==4.1.0

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1

# Code Quality
ruff==0.8.4

# Production Server
gunicorn==23.0.0

# Utilities
python-dateutil==2.9.0

